# src/generated/openapi/api@v1.yaml
openapi: 3.0.3
info:
  title: Wallet Service API v1
  description: Securely provision and manage wallets
  version: 1.0.0
servers:
  - url: http://localhost:8008
    description: Local
paths:
  /api/v1/actions/InitKeygen:
    post:
      operationId: InitKeygen
      summary: Init Keygen
      description: This initializes the keygen process
      x-eov-operation-handler: actions/InitKeygen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitKeygenRequest'
      responses:
        '200':
          description: Successfully initialized keygen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitKeygenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/actions/CreateRoom:
    post:
      operationId: CreateRoom
      summary: Create Room
      description: This creates a room for MPC key generation
      x-eov-operation-handler: actions/CreateRoom
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '200':
          description: Successfully created room
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRoomResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/actions/CreateWalletAccount:
    post:
      operationId: CreateWalletAccount
      summary: Create Wallet Account
      description: This creates a wallet for a given user
      x-eov-operation-handler: actions/CreateWalletAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWalletAccountRequest'
      responses:
        '200':
          description: Successfully created wallet account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateWalletAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/actions/ExportWalletAccount:
    post:
      operationId: ExportWalletAccount
      summary: Export wallet account
      description: This exports a wallet account
      x-eov-operation-handler: actions/ExportWalletAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportWalletAccountRequest'
      responses:
        '201':
          description: Exported wallet account
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/actions/SignMessage:
    post:
      operationId: SignMessage
      summary: Sign message
      description: This signs a message
      x-eov-operation-handler: actions/SignMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignMessageRequest'
      responses:
        '200':
          description: Signed message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/actions/HealthCheck:
    get:
      operationId: HealthCheck
      summary: Service Health Check
      description: Checks service health and availability
      x-eov-operation-handler: actions/HealthCheck
      responses:
        '200':
          description: Successfully revoked session key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
  schemas:
    Chain:
      type: string
      enum:
        - EVM
        - SOL
        - COSMOS
        - BTC
        - FLOW
        - SUI
    UUID:
      type: string
      pattern: ^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$
      minLength: 36
      maxLength: 36
      example: 95b11417-f18f-457f-8804-68e361f9164f
    InitKeygenRequest:
      type: object
      required:
        - chain
        - environmentId
        - userId
      properties:
        chain:
          $ref: '#/components/schemas/Chain'
        environmentId:
          $ref: '#/components/schemas/UUID'
        userId:
          $ref: '#/components/schemas/UUID'
    EncryptedEvervaultString:
      type: string
      pattern: ^ev.*
      example: ev:debug:Tk9D:2X9ZSLTbX7/v+gnZ:AjX3AoXEr7zyV8lMdeBoyZwpoyJ6SedJlSzmdwiOCdSZ:KkpgDb/MC+EarhwFeXwjSjJX25sGgYsvQ0yKSlyNq6UgBiAC6r88f/blx0xVnjA5BimHIKa+WCQLe6bCvcn3m36jI6w/OtJpxCHjIpbdRwdOMa/tPe/YGpZZv4F80ja+Lj9QekDgCF/C1ES+20JSg3n5Wgnkp6IVwH22NiCXvYbdmFMSbUdpH6i42g1AGtLF61nHPlb3Aue5N55b9WS2Xtydnwn6VdSSOxVt1myFYv8Y+tNmnE0UX+99WH8cEMeGFG7s1LqWBDyyf6Ma9MjcENLK1xJKxXcMEuMnigAaMPVnWy/rrVpLv2EbpiWZZqOsVCh0I1e0V6ScRlGIplLtF/14hSJLvwO6F04yptnG31YJUnXCLBuB48lOHexHJZv5ujnGjO1f+69F4q6Z5x++P5d71azu7pAxNNqOCfDmqVOraRMiW+RI2xzLah3LVNsuz3BSVMqGR5UKnEG4fdAuIUz+erWr5lGJdSIwF4JNkAEk2krffHRTlIg66lH6vukM5Kva4SaRsT9ftAthLnBSbv5t:$
    InitKeygenResponse:
      type: object
      required:
        - roomId
      properties:
        eac:
          $ref: '#/components/schemas/EncryptedEvervaultString'
        roomId:
          type: string
        serverKeygenId:
          type: string
    BadRequest:
      type: object
      properties:
        error_code:
          type: string
          enum:
            - bad_request
        error_message:
          type: string
    Forbidden:
      type: object
      required:
        - error_code
        - error_message
      properties:
        error_code:
          type: string
          enum:
            - jwt_invalid
            - jwt_not_authorized
        error_message:
          type: string
    InternalServerError:
      type: object
      properties:
        error_code:
          type: string
          enum:
            - internal_server_error
            - turnkey_api_error
        error_message:
          type: string
          example: Internal Server Error
    CreateRoomRequest:
      type: object
      required:
        - chain
      properties:
        chain:
          $ref: '#/components/schemas/Chain'
    CreateRoomResponse:
      type: object
      required:
        - roomId
      properties:
        roomId:
          type: string
    PartialEAC:
      type: object
      required:
        - userId
        - environmentId
        - serverKeygenInitResult
        - chain
      properties:
        userId:
          $ref: '#/components/schemas/UUID'
        environmentId:
          $ref: '#/components/schemas/UUID'
        serverKeygenInitResult:
          type: string
        chain:
          $ref: '#/components/schemas/Chain'
    CreateWalletAccountRequest:
      type: object
      required:
        - eac
        - roomId
        - clientKeygenId
        - clientBackupKeygenId
      properties:
        eac:
          $ref: '#/components/schemas/PartialEAC'
        roomId:
          type: string
        clientKeygenId:
          type: string
        clientBackupKeygenId:
          type: string
    CreateWalletAccountResponse:
      type: object
      required:
        - eac
        - userId
        - environmentId
        - accountAddress
        - uncompressedPublicKey
      properties:
        eac:
          $ref: '#/components/schemas/EncryptedEvervaultString'
        userId:
          $ref: '#/components/schemas/UUID'
        environmentId:
          $ref: '#/components/schemas/UUID'
        accountAddress:
          type: string
        uncompressedPublicKey:
          type: string
        compressedPublicKey:
          type: string
    EAC:
      type: object
      required:
        - userId
        - uncompressedPublicKey
        - accountAddress
        - environmentId
        - serverKeygenInitResult
        - chain
      properties:
        userId:
          $ref: '#/components/schemas/UUID'
        compressedPublicKey:
          type: string
          pattern: ^[0-9a-fA-F]{33}$
        uncompressedPublicKey:
          type: string
        accountAddress:
          type: string
        environmentId:
          $ref: '#/components/schemas/UUID'
        serverKeygenInitResult:
          type: string
        chain:
          $ref: '#/components/schemas/Chain'
    ExportWalletAccountRequest:
      type: object
      required:
        - eac
        - exportId
        - roomId
      properties:
        eac:
          $ref: '#/components/schemas/EAC'
        exportId:
          type: string
        roomId:
          type: string
    JWT:
      type: string
      pattern: ^[A-Za-z0-9-_=]+\.[A-Za-z0-9-_=]+\.?[A-Za-z0-9-_.+/=]*$
    SignMessageRequest:
      type: object
      required:
        - roomId
        - message
        - eac
      properties:
        jwt:
          $ref: '#/components/schemas/JWT'
        message:
          type: string
        roomId:
          type: string
        eac:
          $ref: '#/components/schemas/EAC'
    SignMessageResponse:
      type: object
      required:
        - signature
      properties:
        signature:
          type: string
    HealthCheckResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequest'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Forbidden'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'